Modal working around. Remove unused addCountry
