+++ src/components/contribs/AreaTrigger.tsx
@@ -1,12 +1,12 @@
-import { MobileDialog, DialogContent, DialogTrigger } from '../ui/MobileScreen'
+import { MobileDialog, DialogContent, DialogTrigger } from '../ui/MobileDialog'
 import AddChildAreaForm, { ChildAreaBaseProps } from './AddChildAreaForm'
 
 export default function AreaTrigger (props: ChildAreaBaseProps): JSX.Element {
   return (
     <div>
-      <MobileDialog>
-        <DialogTrigger asChild>
-          <button type='button' className='btn btn-primary btn-xs'>Edit</button>
+      <MobileDialog modal>
+        <DialogTrigger className='btn btn-accent btn-xs'>
+          Add new
         </DialogTrigger>
         <DialogContent title='Add new child area'>
           <AddChildAreaForm {...props} />
